Update dependencies
🚧 Start refactoring game models tests

Weird error with the same code in adhoc event generation

Finish refactoring game model tests

🎨 Refactor game form tests

🎨 Refactor game task tests

🎨 Refactor calendar feed tests

🎨 Refactor scheduling utils tests

🎨 Refactor game receiver tests

🚧 Trying to refactor game view tests

🚧 Refactor all application and game posting view tests

🎨 Refactor session manipulation view tests

🎨 Finish refactoring games tests

Fixes #132

+++ looking_for_group/games/receivers.py
@@ -277,7 +277,7 @@ def generate_master_event_occurrence_for_adhoc_session(
             master_event = models.GameEvent.objects.create(
                 start=instance.scheduled_time,
                 end=instance.scheduled_time
-                + timedelta(minutes=instance.game.session_length * 60),
+                + timedelta(minutes=int(instance.game.session_length * 60)),
                 title="Ad hoc session for {}".format(instance.game.title),
                 description=instance.game.game_description,
                 creator=instance.game.gm.user,
